\name{"add<-"}
\alias{add<-}
\title{Add new scale}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Add new scale to list.
}
\usage{"add<-"(x, value)}
\arguments{
\item{x}{}
\item{value}{}
}

\details{Will overwrite any existing scales that use the same
output variables.}

\examples{}
\keyword{hplot}
\keyword{internal}
