\name{add_position}
\alias{add_position}
\title{Add position scales.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Add postiion scales to a plot
}
\usage{add_position(plot, pre, position)}
\arguments{
\item{plot}{plot object}
\item{pre}{precompution, see code in \code{\link{ggplot_plot}}}
\item{position}{position ("x" or "y")}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
