\name{aesdefaults}
\alias{aesdefaults}
\title{Aesthetic defaults}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenience method for setting aesthetic defaults
}
\usage{aesdefaults(x, y, ...)}
\arguments{
\item{x}{values from aesthetic mappings}
\item{y}{defaults}
\item{...}{user specified values}
}

\details{}

\examples{}
\keyword{internal}
