% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{alpha}
\alias{alpha}
\title{Give a colour an alpha level}
\usage{
alpha(colour, alpha)
}
\arguments{
\item{colour}{colour}

\item{alpha}{alpha level [0,1]}
}
\description{
Give a colour an alpha level
}
\keyword{internal}

