\name{best_scale}
\alias{best_scale}
\title{Best scale}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Using \code{\link{possible_scales}} and \code{\link{scale_quality}} find
}
\usage{best_scale(data, expand=0)}
\arguments{
\item{data}{vector of data}
\item{expand}{multiplicative expansion factor (experimental)}
}

\details{}

\examples{}
\keyword{internal}
