\name{blueyellow2}
\alias{blueyellow2}
\title{Create blue-yellow colour scheme}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create a colour scheme that varies continuousy along one ray of a colour space
}
\usage{blueyellow2(n)}
\arguments{
\item{n}{number of colours to produce}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
