\name{breaks}
\alias{breaks}
\alias{breaks.categorical}
\alias{breaks.continuous}
\alias{breaks.gradient}
\alias{breaks.ps_double}
\alias{breaks.ps_map}
\alias{breaks.manual}
\title{Scale breaks}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Compute breaks of scale object
}
\usage{breaks(scale, ...)}
\arguments{
\item{scale}{scale object}
\item{...}{...}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
