\name{brewer_palettes}
\alias{brewer_palettes}
\title{Get Brewer colour palettes}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenience function to retrieve private RColorBrewer palettes.
}
\usage{brewer_palettes(type)}
\arguments{
\item{type}{type of palettes to retrieve}
}

\details{}

\examples{}
\keyword{internal}
