\name{brewer_type}
\alias{brewer_type}
\title{Brewer type}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Return the type of factor in Cynthia brewers scheme.
}
\usage{brewer_type(x)}
\arguments{
\item{x}{}
}
\value{character string giving Brewer type}
\details{}

\examples{}
\keyword{internal}
