% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetic-maps.r
\name{chop}
\alias{chop}
\title{Chop a continuous variable into a categorical variable.}
\usage{
chop(x, n = 5, method = "quantiles", midpoint = 0, digits = 2)
}
\arguments{
\item{x}{continuous variable to chop into pieces}

\item{n}{number of bins to chop into}

\item{method}{method to use: quantiles (approximately equal numbers), cut (equal lengths) or pretty}

\item{midpoint}{mid point for diverging factors}

\item{digits}{number of digits to show in labels}
}
\description{
Chop provides a convenient interface to the main methods of
converting a continuous variable into a categorical variable.
}

