% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetic-maps.r
\name{chop_auto}
\alias{chop_auto}
\title{Automatic chop
Keep categorical variables as is, chop up continuous variable}
\usage{
chop_auto(x)
}
\arguments{
\item{x}{data}
}
\description{
Automatic chop
Keep categorical variables as is, chop up continuous variable
}
