\name{defaultaesthetics}
\alias{defaultaesthetics}
\title{Default aesthetics}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Modify the default aesthetics for a plot
}
\usage{defaultaesthetics(plot, aesthetics)}
\arguments{
\item{plot}{plot object}
\item{aesthetics}{new list of aesthetics}
}
\value{modified plot object}
\details{}

\examples{}
\keyword{hplot}
