% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{defaultaesthetics}
\alias{defaultaesthetics}
\title{Modify the default aesthetics for a plot}
\usage{
defaultaesthetics(plot, aesthetics)
}
\arguments{
\item{plot}{plot object}

\item{aesthetics}{new list of aesthetics}
}
\value{
modified plot object
}
\description{
Modify the default aesthetics for a plot
}

