\name{defaultgrob}
\alias{defaultgrob}
\alias{deafultgrob.scale}
\alias{defaultgrob.categorical}
\alias{defaultgrob.gradient}
\alias{defaultgrob.default}
\title{Deafult grob}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Return the default grob to use for creating a legend
}
\usage{defaultgrob(x)}
\arguments{
\item{x}{scale object}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
