% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.r
\name{expand_range}
\alias{expand_range}
\title{Expand range
Convenience function for expanding a range with a multiplicative
or additive constant.}
\usage{
expand_range(range, mul = 0, add = 0)
}
\arguments{
\item{range}{range of data}

\item{mul}{multiplicative constract}

\item{add}{additive constant}
}
\description{
Expand range
Convenience function for expanding a range with a multiplicative
or additive constant.
}
