% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.r
\docType{data}
\name{theme_bw}
\alias{theme_bw}
\alias{theme_default}
\alias{ggopt}
\alias{ggtheme}
\title{Set global options for ggplot.}
\format{An object of class \code{list} of length 2.}
\usage{
theme_bw

theme_default

ggopt(...)

ggtheme(theme)
}
\arguments{
\item{...}{named options to set}

\item{theme}{list of options set}
}
\description{
These are aliased into every plot object, so that \code{p$grid.col} will
return the default grid colour, unless it has been overriden for a particular
plot object.  You can change the global options using the function, or the
options for a specific plot by setting the values directly on the object.  See
the examples for more details.
}
\details{
Colour settings:

\itemize{
	\item axis.colour: axis text and line colour ("black")
	\item background.colour: background text colour ("black"), used for title
	\item background.fill:   background fill ("white")
	\item grid.colour: plot grid colour ("white")
	\item grid.fill:   plot grid background fill ("grey90")
}

Strip settings

\itemize{
	\item strip.text:   function with two arguments (variable, and value) used for
		generating strip labels
	\item strip.gp: graphic parameter settings for the strip
	\item strip.text.gp:  graphic parameter settings for the strip text
}

Legend settings

\itemize{
	\item legend.position:   position of legend: "none" to hide legend;
		"left", "right", "top", "bottom", for positioning outside of plot;
		c(x, y) for positioning on top of plot
}

Other settings:

\itemize{
	\item aspect.ratio: aspect ratio of facets.  Set to \code{NULL} to allow
			 to vary with device size
}
}
\examples{
# Change global options
ggopt(background.fill = "black", background.color ="white")
p <- ggpoint(ggplot(reshape::tips, smoker ~ sex,aesthetics = list(y = tip, x = total_bill)))
p

# Change individual plot options
p$background.fill = "white"
p
p$strip.text.gp <- grid::gpar(col="red", fontsize=8)
p$background.colour <- "pink"
p$grid.colour <- "green"
p$grid.fill <- "blue"
p # a very ugly plot!

# Use ggtheme(theme_default) to reset back to the
# default theme.
ggtheme(theme_default)
}
\keyword{datasets}
