\name{ggpretty}
\alias{ggpretty}
\title{GG Pretty}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Pretty axis breaks
}
\usage{ggpretty(x)}
\arguments{
\item{x}{values to prettify}
}

\details{Same as \code{\link{grid.pretty}} but contains minimum and
maximum of data as well.  Useful for legends.}

\examples{}
\keyword{internal}
