% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs-complex.r
\name{ggquantile}
\alias{ggquantile}
\title{Grob function: quantiles}
\usage{
ggquantile(plot, aesthetics = list(), ..., data = NULL)
}
\arguments{
\item{plot}{the plot object to modify}

\item{aesthetics}{named list of aesthetic mappings, see details for more information}

\item{...}{other options, see details for more information}

\item{data}{data source, if not specified the plot default will be used}
}
\description{
This can be used a continuous analogue of a boxplot (see \code{\link{ggboxplot}})
Lines will be automatically sized to reflect their distance from the median.
}
\details{
Aesthetic mappings that this grob function understands:

\itemize{
  \item \code{x}:x position (required)
  \item \code{y}:y position (required)
  \item \code{weight}: observation weights
}

These can be specified in the plot defaults (see \code{\link{ggplot}}) or
in the \code{aesthetics} argument.  If you want to modify the position
of the points or any axis options, you will need to add a position scale to
the plot.  These functions start with \code{ps}, eg.
\code{\link{pscontinuous}} or \code{\link{pscategorical}}

Other options:

\itemize{
  \item \code{quantiles}:quantiles to display
  \item \code{formula}:formula to use in quantile regression
}
}
\examples{
\dontrun{
library(ggplot2movies)
m <- ggplot(movies, aesthetics=list(y=length, x=rating))
ggquantile(m)
}
}
\seealso{
\code{\link[quantreg]{rq}} for the code used to fit the quantile regression
}
