\name{grob_grid}
\alias{grob_grid}
\title{Grob grid}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Build up a subtle background grid
}
\usage{grob_grid(aesthetics, xbreaks, ybreaks, fill=ggopt()$grid.fill, colour=ggopt()$grid.colour, ...)}
\arguments{
\item{aesthetics}{not used}
\item{xbreaks}{x axis lines}
\item{ybreaks}{y axis lines}
\item{fill}{not used}
\item{colour}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
