\name{grobPromise}
\alias{grobPromise}
\title{Create a new grob promise object}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{grobPromise(name, data, params, aesthetics)}
\arguments{
\item{name}{}
\item{data}{}
\item{params}{}
\item{aesthetics}{}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
