\name{guides.default}
\alias{guides.default}
\title{Default guides}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Construct a default guide (legend) for a scale
}
\usage{guides.default(scale, ...)}
\arguments{
\item{scale}{scale}
\item{...}{}
}

\details{This is used for automatic legends.}

\examples{}
\keyword{hplot}
