\name{guides}
\alias{guides}
\alias{guides.ps_double}
\alias{guides.ps_map}
\alias{guides.scale_cont_colour}
\title{Guides}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create guides for the given scale
}
\usage{guides(scale, ...)}
\arguments{
\item{scale}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
