\name{guides.position}
\alias{guides.position}
\title{Position guides}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create x or y axis depending on variable name.
}
\usage{guides.position(scale, ...)}
\arguments{
\item{scale}{scale}
\item{...}{not currently used}
}

\details{}

\examples{}
\keyword{hplot}
