\name{input}
\alias{input}
\alias{input.scale}
\alias{input.ps_double}
\alias{input.ps_map}
\alias{input.scale_cont_colour}
\title{Input}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Return what input variable this scale uses
}
\usage{input(scale)}
\arguments{
\item{scale}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
