\name{map_aesthetic}
\alias{map_aesthetic}
\alias{map_aesthetic.scale}
\alias{map_aesthetic.categorical}
\alias{map_aesthetic.continuous}
\alias{map_aesthetic.gradient}
\alias{map_aesthetic.ps_double}
\alias{map_aesthetic.ps_map}
\alias{map_aesthetic.scale_cont_colour}
\alias{map_aesthetic.manual}
\title{Map}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Map raw data to aesthetics using the provided scale
}
\usage{map_aesthetic(scale, data, ...)}
\arguments{
\item{scale}{}
\item{data}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
