\name{map_aesthetic.scales}
\alias{map_aesthetic.scales}
\title{Map scales.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Applies scales to data to return a data frame
}
\usage{map_aesthetic.scales(scale, data, ...)}
\arguments{
\item{scale}{scale}
\item{data}{data}
\item{...}{other arguments (unused)}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
