\name{map_all}
\alias{map_all}
\title{Map all}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Map all grobs with scale
}
\usage{map_all(scale, matrix)}
\arguments{
\item{scale}{scale to map with}
\item{matrix}{matrix of grobs}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
