\name{map_colour_gradient}
\alias{map_colour_gradient}
\alias{map_color_gradient}
\title{Aesthetic mapping: colour gradient}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Map values to a colour gradient
}
\usage{map_colour_gradient(x, low="red", mid="white",high="black", midpoint = 0, from=range(x, na.rm=TRUE))}
\arguments{
\item{x}{data vector}
\item{low}{colour to use at bottom of scale}
\item{mid}{colour to use at middle of scale}
\item{high}{colour to use at top of scale}
\item{midpoint}{where mid point colour should be used}
\item{from}{}
}

\details{}

\examples{}
\keyword{hplot}
