\name{map_colour_hsv}
\alias{map_colour_hsv}
\alias{map_color_hsv}
\title{Aesthetic mapping: hsv components of colour}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Map variables to hue, saturation or value
}
\usage{map_colour_hsv(h=1, s=1, v=1, a=1, h.to=c(0,1), s.to=c(0,1), v.to=c(0,1), a.to=c(0,1), h.from=range(h, na.rm=TRUE), s.from = range(s, na.rm=TRUE), v.from = range(v, na.rm=TRUE), a.from = range(a, na.rm=TRUE))}
\arguments{
\item{h}{hue}
\item{s}{saturation}
\item{v}{value}
\item{a}{alpha}
\item{h.to}{hue range}
\item{s.to}{saturation range}
\item{v.to}{value range}
\item{a.to}{alpha range}
\item{h.from}{}
\item{s.from}{}
\item{v.from}{}
\item{a.from}{}
}

\details{}

\examples{}
\keyword{hplot}
