\name{.map_colour}
\alias{.map_colour}
\title{Map colour}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenience function to power \code{\link{map_colour_hsv}},
}
\usage{.map_colour(colours, tos, froms, colour_function)}
\arguments{
\item{colours}{list of colour vectors}
\item{tos}{list of colour tos in same order as colours}
\item{froms}{function to produce colours in \#rrggbbaa form}
\item{colour_function}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
