\name{map_colour_rgb}
\alias{map_colour_rgb}
\alias{map_color_rgb}
\title{Aesthetic mapping: rgb components of colour}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Map variables to red, green or blue components.
}
\usage{map_colour_rgb(r=0, g=0, b=0, a=1, r.to = c(0,1), g.to=c(0,1), b.to=c(0,1), a.to=c(0,1), r.from = range(r, na.rm=TRUE), g.from = range(g, na.rm=TRUE), b.from = range(b, na.rm=TRUE), a.from = range(a, na.rm=TRUE))}
\arguments{
\item{r}{red}
\item{g}{green}
\item{b}{blue}
\item{a}{alpha}
\item{r.to}{red to}
\item{g.to}{green to}
\item{b.to}{blue to}
\item{a.to}{alpha to}
\item{r.from}{}
\item{g.from}{}
\item{b.from}{}
\item{a.from}{}
}

\details{}

\examples{}
\keyword{hplot}
