\name{map_linetype}
\alias{map_linetype}
\title{Aesthetic mapping: line type}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Map values to line types
}
\usage{map_linetype(x)}
\arguments{
\item{x}{data vector}
}

\details{If x is not a factor, will be converted to one by \code{link{chop_auto}}.
Can display at most 4 different categories.}

\examples{}
\keyword{manip}
