\name{output}
\alias{output}
\alias{output.scale}
\alias{output.ps_double}
\alias{output.ps_map}
\alias{output.scale_cont_colour}
\title{Output}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Return what output variables this scale produces
}
\usage{output(scale)}
\arguments{
\item{scale}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
