\name{plot_add_grobs}
\alias{plot_add_grobs}
\title{Add grobs to plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Add grobs to plot grob list
}
\usage{plot_add_grobs(plot, grob_matrix)}
\arguments{
\item{plot}{plot object}
\item{grob_matrix}{matrix of grobs}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
