\name{position_adjust}
\alias{position_adjust}
\title{Position adjust}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This function extracts out some of the code necessary to perform
}
\usage{position_adjust(aesthetics, avoid, direction, adjust=1)}
\arguments{
\item{aesthetics}{list of aesthetics}
\item{avoid}{avoid method to use}
\item{direction}{direction (currently only works with vertical)}
\item{adjust}{adjustment factor if multiple rows correspond to same grob (awful hack)}
}

\details{}

\examples{}
\keyword{internal}
