\name{position_apply}
\alias{position_apply}
\title{Position apply}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Apply a function to x and y position scales.
}
\usage{position_apply(scales, f, ...)}
\arguments{
\item{scales}{scales}
\item{f}{function to apply}
\item{...}{other arguments to pass to f}
}

\details{This is a convience method because position scales can
be made up of two separate scales, or one scale that provides
both x and y position mappings.}

\examples{}
\keyword{hplot}
\keyword{internal}
