\name{possible_scales}
\alias{possible_scales}
\title{Possible scales}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
For a given data set, create all possible scales from the
}
\usage{possible_scales(data, expand=0)}
\arguments{
\item{data}{vector of data}
\item{expand}{multiplicative expansion factor (experimental)}
}

\details{}

\examples{}
\keyword{internal}
