\name{preprocess}
\alias{preprocess}
\title{Preprocess}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Precprocess panel given data set using information in grobPromise
}
\usage{preprocess(gp, data)}
\arguments{
\item{gp}{grobPromise to preprocess}
\item{data}{}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
