\name{print.categorical}
\alias{print.categorical}
\title{Print categorical details}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Print moderately useful details of this categorical scale.
}
\usage{print.categorical(x, ...)}
\arguments{
\item{x}{scale object}
\item{...}{not used}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
