\name{print.manual}
\alias{print.manual}
\title{Print manual details}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Print moderately useful details of this manual scale.
}
\usage{print.manual(x, ...)}
\arguments{
\item{x}{scale object}
\item{...}{not used}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
