\name{print.scale}
\alias{print.scale}
\title{Print scale}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Print a moderately useful description of a scale object
}
\usage{print.scale(x, ...)}
\arguments{
\item{x}{scale object}
\item{...}{unused}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
