\name{ps_map}
\alias{ps_map}
\title{Map projection scale}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Map projections
}
\usage{ps_map(projection="mercator", params=NULL)}
\arguments{
\item{projection}{projection to use, see \code{\link[mapproj]{mapproject}} for possible values}
\item{params}{list of parameters passed to \code{\link[mapproj]{mapproject}}}
}

\details{This allows you to use map type projection.}

\examples{}
\keyword{hplot}
