% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.r
\name{ps_map}
\alias{ps_map}
\title{Map projection scale}
\usage{
ps_map(projection = "mercator", params = NULL)
}
\arguments{
\item{projection}{projection to use, see \code{\link[mapproj]{mapproject}} for possible values}

\item{params}{list of parameters passed to \code{\link[mapproj]{mapproject}}}
}
\description{
This allows you to use map type projection.
}
