\name{pstime}
\alias{pstime}
\alias{position_time}
\alias{breaks.time}
\alias{labels.time}
\alias{range.time}
\alias{time_scale}
\title{Time scale takes standard numeric scale (in seconds since epoch)}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
and displays nice time intervals instead
}
\usage{pstime(plot = .PLOT, variable="x", name="", transform=trans_none, range=c(NA,NA), expand=c(0.05, 0), by="weeks")}
\arguments{
\item{plot}{}
\item{variable}{}
\item{name}{}
\item{transform}{}
\item{range}{}
\item{expand}{}
\item{by}{}
}

\details{}

\examples{}
\keyword{internal}
