\name{sc}
\alias{sc}
\alias{defaultgrob.manual}
\alias{guides.manual}
\alias{labels.categorical}
\alias{labels.continuous}
\alias{labels.gradient}
\alias{labels.manual}
\alias{labels.ps_double}
\alias{labels.ps_map}
\alias{position_categorical}
\alias{position_continuous}
\alias{print.continuous}
\alias{print.gradient}
\alias{range.categorical}
\alias{range.continuous}
\alias{range.ps_double}
\alias{scale_colour}
\alias{scale_fill}
\alias{scale_fill_brewer}
\alias{scale_fill_hcl}
\alias{scale_fill_hsv}
\alias{scale_fill_rgb}
\alias{scale_gradient}
\alias{scale_hcl}
\alias{scale_hsv}
\alias{scale_linetype}
\alias{scale_manual}
\alias{scale_rgb}
\alias{scale_shape}
\alias{scale_size}
\title{Scale methods}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{sc()}
\arguments{

}

\details{}

\examples{}
\keyword{internal}
