\name{scale_apply_combine_map}
\alias{scale_apply_combine_map}
\title{Scale apply combine map}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{scale_apply_combine_map(scale, rdmatrix)}
\arguments{
\item{scale}{}
\item{rdmatrix}{}
}

\details{}

\examples{scale_new(c(size="3 * height"), "size", range=c(0, 5), class="size")
scale_new(c(x="weight", y="height"), c("x", "y"), class=c("equal", "position"))
scale_new(c(group="age"), c("colour", "glyph"), class="group")
scale_new(c(h="age"), c("colour"), class="hsv", range=list(l=c(4,4), h=c(30,40)))
scale_new(c())}
\keyword{internal}
