\name{scale_apply_to_data}
\alias{scale_apply_to_data}
\title{Given a scale and a data.frame, this runs the expressions}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
in the input and returns a list
}
\usage{scale_apply_to_data(scale, data=data.frame())}
\arguments{
\item{scale}{}
\item{data}{}
}

\details{}

\examples{}
\keyword{internal}
