\name{scale_cont_colour}
\alias{scale_cont_colour}
\title{Scale: colour (continuous)}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Scale colour along a continuous path
}
\usage{scale_cont_colour(inputs, name="", variable="colour", to=list())}
\arguments{
\item{inputs}{}
\item{name}{}
\item{variable}{}
\item{to}{}
}

\details{This scale class is the workhorse behind:

\item \code{\link{scrgb}}
\item \code{\link{schsv}}
\item \code{\link{schcl}}

See those function for more details.}

\examples{}
\keyword{hplot}
\keyword{internal}
