\name{scale_map}
\alias{scale_map}
\title{@arguments scale object}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
@arguments \code{data.frame} to be mapped
}
\usage{scale_map(scale, data, domain, range=scale$range)}
\arguments{
\item{scale}{scale object}
\item{data}{\code{data.frame} to be mapped}
\item{domain}{named list of domains (one for each) input variable, each domain being a list of length two, the first element containing the range of the numeric component, the second the text labels of the categorical component}
\item{range}{}
}

\details{}

\examples{}
\keyword{internal}
