\name{scale_mapping}
\alias{scale_mapping}
\title{Scale mapping}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Text string describing how the mapping of the scale works
}
\usage{scale_mapping(x)}
\arguments{
\item{x}{scale object}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
