\name{scale_new}
\alias{scale_new}
\alias{scale_domain}
\alias{scale_map_single}
\alias{scale_desired_size}
\alias{range2}
\title{New scales functionality}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{scale_new(input, output=names(input), name = paste(input, collapse="/"), ..., class=output)}
\arguments{
\item{input}{}
\item{output}{}
\item{name}{}
\item{...}{}
\item{class}{}
}

\details{}

\examples{}
\keyword{internal}
