\name{scale_quality}
\alias{scale_quality}
\title{Scale quality}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Compute the how good a scale is for a given data set
}
\usage{scale_quality(s, q, data)}
\arguments{
\item{s}{scale values}
\item{q}{unit of Q used to create scale}
\item{data}{vector of data}
}

\details{}

\examples{}
\keyword{internal}
