\name{scales}
\alias{scales}
\alias{input.scales}
\alias{output.scales}
\alias{print.scales}
\alias{range.scales}
\alias{guides.scales}
\title{Scaleslist object}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create a list of scales objects
}
\usage{scales(...)}
\arguments{
\item{...}{scales objects}
}

\details{The scales output maintains a list of scale objects.

\item input and output variables
\item maps a data frames using those scales
\item generates ready to use scales}

\examples{}
\keyword{hplot}
\keyword{internal}
