% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-categorical.r
\name{sccolour}
\alias{sccolor}
\alias{sccolour}
\alias{scfill}
\title{Scale: categorical colour
Create a scale for categorical colours.}
\usage{
sccolour(plot, name = "", h = c(0, 360), l = 65, c = 100, alpha = 1)

sccolor(plot, name = "", h = c(0, 360), l = 65, c = 100, alpha = 1)

scfill(plot, name = "", h = c(0, 360), l = 75, c = 100, alpha = 1)
}
\arguments{
\item{plot}{ggplot object}

\item{name}{name of the scale (used in the legend)}

\item{h}{range of hues to use}

\item{l}{luminance value}

\item{c}{chroma value}

\item{alpha}{alpha value}
}
\description{
Continuous variables will automatically be converted to categorical
using \code{\link{chop_auto}}.  You may want to use \code{\link{chop}}
to convert the values yourself for finer control.
}
\details{
This scale is automatically added when you have colour in your list of
aesthetics.  For finer control, you will need to set the scale
yourself.  See the example for some ideas.
}
\examples{
library(ggplot2movies)
p <- ggplot(movies, aes=list(x=mpaa, y=rating))
ggjitter(p, list(colour=rating))
ggjitter(p, list(colour=length))
ggjitter(p, list(colour=chop(length)))
ggjitter(p, list(colour=chop(length,3)))
sccolour(ggjitter(p, list(colour=chop(length,3))), 2)
}

