\name{scfillbrewer}
\alias{scfillbrewer}
\title{Scale: Brewer colours}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Use Brewer colour scheme for colour fill.
}
\usage{scfillbrewer(plot = .PLOT, name="", palette=1)}
\arguments{
\item{plot}{plot to add scale to}
\item{name}{name of the scale (used in the legend)}
\item{palette}{Color Brewer palette to use, see \code{\link[RColorBrewer]{brewer.pal}} for details.  Note that palette type is chosen automatically.}
}

\details{}

\examples{}
\keyword{hplot}
