% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-categorical.r
\name{scfillbrewer}
\alias{scfillbrewer}
\title{Scale: Brewer colours
Use Brewer colour scheme for colour fill.}
\usage{
scfillbrewer(plot, name = "", palette = 1)
}
\arguments{
\item{plot}{ggplot object}

\item{name}{name of the scale (used in the legend)}

\item{palette}{Color Brewer palette to use, see
\code{\link[RColorBrewer]{brewer.pal}} for details. Note that palette type
is chosen automatically.}
}
\description{
Scale: Brewer colours
Use Brewer colour scheme for colour fill.
}

