% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous.r
\name{scgradient}
\alias{scgradient}
\alias{scfillgradient}
\title{Scale: colour gradient
Scale a continuous variable along a colour gradient.}
\usage{
scgradient(
  plot,
  name = "",
  low = "red",
  mid = "white",
  high = "black",
  midpoint = 0,
  range = c(NA, NA)
)

scfillgradient(
  plot,
  name = "",
  low = "red",
  mid = "white",
  high = "black",
  midpoint = 0,
  range = c(NA, NA)
)
}
\arguments{
\item{plot}{plot}

\item{name}{namen of the scale (used in the legend)}

\item{low}{colour at low end of scale}

\item{mid}{colour at middle of scale}

\item{high}{colour at top of scale}

\item{midpoint}{definition of midpoint}

\item{range}{range to scale data to}
}
\description{
This scale creates a continuous colour gradient from the
low colour to the mid colour to high colour, as defined in the
arguments.
}
\examples{
p <- scgradient(ggplot(reshape::tips, aes = list(x = sex, y = day)))
ggjitter(p, list(colour = total_bill))
ggjitter(p, list(colour = tip))

p <- ggjitter(p, list(colour = tip))
scgradient(p, low="yellow")
scgradient(p, high="green", midpoint=5)
}
