% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-cont-colour.r
\name{schcl}
\alias{scfillhcl}
\alias{schcl}
\title{Scale: colour (hcl)
Scale continuous variables to hue, chroma and luminance components of colour}
\usage{
schcl(plot, name = "", to = list())

scfillhcl(plot, name = "", to = list())
}
\arguments{
\item{plot}{plot}

\item{name}{namen of the scale (used in the legend)}

\item{to}{named list of target ranges (r.to, g.to, b.to, a.to)}
}
\description{
This colour map is the most perceptually uniform.  However, use multiple
mappings with care.  It is often a good idea to restrict the range of the
hue, as shown in the example.
}
\details{
Note: alpha mappings only work with the Quartz and PDF devices.
}
\examples{
p <- schcl(ggplot(reshape::tips, aes = list(x = total_bill, y = 1)))
ggjitter(p, list(h = tip))
ggjitter(p, list(c = tip))
ggjitter(p, list(l = tip))
}

