% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-categorical.r
\name{sclinetype}
\alias{sclinetype}
\title{Scale: line type}
\usage{
sclinetype(plot = .PLOT, name = "")
}
\arguments{
\item{plot}{ggplot object}

\item{name}{name of the scale (used in the legend)}
}
\description{
This scale is automatically added to the plot when you use the linetype
aesthetic.  As there are no options to this scale, you shouldn't ever
need to add it yourself.
}
\examples{
p <- ggplot(mtcars, aes=list(x=mpg, y=wt, linetype=cyl))
ggline(p)
ggline(sclinetype(p))
}

