% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-manual.r
\name{scmanual}
\alias{scmanual}
\title{Scale: manual}
\usage{
scmanual(plot, variable = "x", name = "", breaks = NULL,
  labels = as.character(breaks), grob = function(x) grob_point(x,
  unique = FALSE))
}
\arguments{
\item{plot}{plot}

\item{variable}{variable ("x" or "y")}

\item{name}{namen of the scale (used in the legend)}

\item{breaks}{set breaks manually}

\item{labels}{character vector of break labels}

\item{grob}{grob function to use when drawing legend}
}
\description{
This scale function allows you complete control over the
scale.
}
\details{
Supply labels and breaks to produce a legend.
}
