% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-categorical.r
\name{scshape}
\alias{scshape}
\title{Scale: shape}
\usage{
scshape(plot, name = "", solid = TRUE)
}
\arguments{
\item{plot}{ggplot object}

\item{name}{name of the scale (used in the legend)}

\item{solid}{should points be solid or hollow?}
}
\description{
This scale is automatically added when you use the shape aesthetic
mapping.  By using this scale you can explicitly decide whether the
points used should be hollow or solid.
}
\examples{
p <- ggplot(mtcars, aes=list(x=mpg, y=wt, shape=cyl))
ggpoint(p)
ggpoint(scshape(p, FALSE))
}
