\name{setdata}
\alias{setdata}
\title{Set default dataset for a plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Set the default data set for a plot object
}
\usage{setdata(p = .PLOT, data)}
\arguments{
\item{p}{plot object, if not specified will use current plot}
\item{data}{new data set}
}

\details{}

\examples{}
\keyword{hplot}
