% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{setfacets}
\alias{setfacets}
\title{Set facetting formula and margins for a plot}
\usage{
setfacets(p = .PLOT, formula = . ~ ., margins = FALSE)
}
\arguments{
\item{p}{plot object, if not specified will use current plot}

\item{formula}{formula describing row and column layout, see
\code{\link[reshape]{cast}} for more details}

\item{margins}{a vector of names giving which margins to display, can include
\code{grand_row} and \code{grand_col} or us TRUE to display all margins}
}
\description{
Set facetting formula and margins for a plot
}

