\name{tailcolor}
\alias{tailcolor}
\title{Tail colour}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Colour scheme that emphasizes differences in tails.
}
\usage{tailcolor(n=100, q1=0.10, q2=0.25, q3 = 1-q2,q4 = 1-q1)}
\arguments{
\item{n}{number of colours to produce}
\item{q1}{quantile for first colour}
\item{q2}{quantile for second colour}
\item{q3}{quantile for third colour, defaults to symmetric}
\item{q4}{quantile for fourth colour, defaults to symmetric}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
