\name{"update<-"}
\alias{update<-}
\alias{update<-.scale}
\alias{update<-.categorical}
\alias{update<-.continuous}
\alias{update<-.gradient}
\alias{update<-.ps_equal}
\alias{update<-.ps_map}
\alias{update<-.scale_cont_colour}
\alias{update<-.manual}
\title{Update}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Update a scale with data values.
}
\usage{"update<-"(x, value)}
\arguments{
\item{x}{}
\item{value}{}
}

\details{This is used to teach each scale about the full range of the data
so that all panels share a common scale.}

\examples{}
\keyword{hplot}
\keyword{internal}
