% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{setdata}
\alias{setdata}
\title{Set the default data set for a plot object}
\usage{
setdata(p, data)
}
\arguments{
\item{p}{plot object}

\item{data}{new data set}
}
\description{
Set the default data set for a plot object
}
